requirements.txt
slack-bolt>=1.18,<2
slack-sdk>=3.21,<4
httpx>=0.25,<1
apscheduler>=3.10,<4
python-dotenv>=1.0,<2
